[FE] RENAME: 계정 연결 카드 컨테이너 파일 및 컴포넌트 이름 변경 #1739
